+++ stdlib/funcs.go
@@ -101,6 +101,7 @@ func Functions(basedir string, experiments []string) map[string]function.Functio
 	tmfuncs["tm_hcldecode"] = HCLDecode()
 
 	tmfuncs["tm_slug"] = SlugFunc()
+	tmfuncs["tm_joinlist"] = JoinListFunc()
 
 	tmfuncs["tm_tree"] = TreeFunc()